requirements.txt
langchain>=0.1.0
langgraph>=0.1.0
langsmith>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-chroma>=0.1.0
chromadb>=0.4.0
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
rarfile>=4.0
dashscope>=1.14.0
langchain-text-splitters>=0.1.0
chardet>=5.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0 